print.adaboost {fastAdaboost}	R Documentation

Print adaboost.m1 model summary

Description

S3 method to print an adaboost object

Usage

## S3 method for class 'adaboost'
print(x, ...)

Arguments

x
object of class adaboost

...
arguments passed to print.default

Details

Displays basic information on the model, such as function call, dependent variable, the number of trees, and weights assigned to each tree

Value

None

See Also

print.real_adaboost

Examples

fakedata <- data.frame( X=c(rnorm(100,0,1),rnorm(100,1,1)), Y=c(rep(0,100),rep(1,100) ) )
fakedata$Y <- factor(fakedata$Y)
test_adaboost <- adaboost(Y~X, fakedata, 10)
print(test_adaboost)

[Package fastAdaboost version 1.0.0 Index]