print.real_adaboost {fastAdaboost}	R Documentation

Print real adaboost model summary

Description

S3 method to print a real_adaboost object

Usage

## S3 method for class 'real_adaboost'
print(x, ...)

Arguments

x
object of class real_adaboost

...
arguments passed to print.default

Details

Displays basic information on the model, such as function call, dependent variable and the number of trees

Value

None

See Also

print.adaboost

Examples

fakedata <- data.frame( X=c(rnorm(100,0,1),rnorm(100,1,1)), Y=c(rep(0,100),rep(1,100) ) )
fakedata$Y <- factor(fakedata$Y)
test_real_adaboost<- real_adaboost(Y~X, fakedata, 10)
print(test_real_adaboost)

[Package fastAdaboost version 1.0.0 Index]