real_adaboost {fastAdaboost}	R Documentation

Real Adaboost algorithm

Description

Implements Zhu et al's real adaboost or SAMME.R algorithm

Usage

real_adaboost(formula, data, nIter, ...)

Arguments

formula
Formula for models

data
Input dataframe

nIter
no. of classifiers

...
other optional arguments, not implemented now

Details

This implements the real adaboost algorithm for a binary classification task. The target variable must be a factor with exactly two levels. The final classifier is a linear combination of weak decision tree classifiers. Real adaboost uses the class probabilities of the weak classifiers to iteratively update example weights. It has been found to have lower generalization errors than adaboost.m1 for the same number of iterations.

Value

object of class real_adaboost

References

Zhu, Ji, et al. “Multi-class adaboost” Ann Arbor 1001.48109 (2006): 1612.

See Also

adaboost,predict.real_adaboost

Examples

fakedata <- data.frame( X=c(rnorm(100,0,1),rnorm(100,1,1)), Y=c(rep(0,100),rep(1,100) ) )
fakedata$Y <- factor(fakedata$Y)
test_adaboost <- real_adaboost(Y~X, data=fakedata,10)

[Package fastAdaboost version 1.0.0 Index]